requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
langchain-ollama
langchain-text-splitters
unstructured
chromadb
langchain-chroma
langchain-classic
langchain-core
faster-whisper
pyaudio
pvrecorder
pvporcupine
requests